backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-pptx==1.0.2
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
aiofiles==24.1.0
Pillow==10.4.0
aspose-slides==24.8.0